+++ src/feecc_workbench/_Barcode.py
