+++ src/feecc_workbench/_image_generation.py
