+++ tests/test_api.py
@@ -3,7 +3,7 @@ import time
 import requests
 from fastapi.testclient import TestClient
 
-from feecc_hub_src.app import api
+from src.app import api
 
 client = TestClient(api)
